backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2